refactor: split design-system page into section components
